+++ R/cbind.R
@@ -45,9 +45,8 @@ cbind.mids <- function(x, y = NULL, ...) {
   blocknames <- make.unique(xynames)
   names(blocknames) <- xynames
   names(blocks) <- blocknames
-  ct <- c(attr(x$blocks, "calltype"), rep("pred", ncol(y)))
-  names(ct) <- blocknames
-  attr(blocks, "calltype") <- ct
+  modeltype <- c(x$modeltype, rep("pred", ncol(y)))
+  names(modeltype) <- blocknames
 
   m <- x$m
 
@@ -123,6 +122,7 @@ cbind.mids <- function(x, y = NULL, ...) {
     predictorMatrix = predictorMatrix,
     visitSequence = visitSequence,
     formulas = formulas,
+    modeltype = modeltype,
     post = post,
     blots = blots,
     ignore = ignore,
@@ -173,9 +173,8 @@ cbind.mids.mids <- function(x, y, call) {
   blocknames <- make.unique(xynames)
   names(blocknames) <- xynames
   names(blocks) <- blocknames
-  ct <- c(attr(xblocks, "calltype"), attr(yblocks, "calltype"))
-  names(ct) <- blocknames
-  attr(blocks, "calltype") <- ct
+  modeltype <- c(x$modeltype, y$modeltype)
+  names(modeltype) <- blocknames
 
   m <- x$m
   nmis <- c(x$nmis, y$nmis)
@@ -294,6 +293,7 @@ cbind.mids.mids <- function(x, y, call) {
     predictorMatrix = predictorMatrix,
     visitSequence = visitSequence,
     formulas = formulas,
+    modeltype = modeltype,
     post = post,
     blots = blots,
     ignore = ignore,